% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSite.R
\name{addSite}
\alias{addSite}
\title{addSite}
\usage{
addSite(park, SiteCode, SiteName, Coordinates, Type)
}
\arguments{
\item{park}{A \code{Park} object}

\item{SiteCode}{A short code to designate the site, stored as a length 1 character \code{vector}.}

\item{SiteName}{A name for the site. Stored as a length 1 character \code{vector}.}

\item{Coordinates}{A length 2 numeric \code{vector} containing the latitude and longitude of the site in WGS84 coordinates.}

\item{Type}{The type of water body at the site - e.g. Stream, River, Lake, Ocean etc. Stored as a length 1 character \code{vector}.}
}
\value{
Returns the park object with the site added.
}
\description{
This function adds a site to an existing \code{Park} object.
}

