% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRNWater_Characteristic_Class_def.R
\docType{class}
\name{Characteristic-class}
\alias{Characteristic-class}
\title{S4 Class Definition for Characteristic object in  NCRNWater}
\description{
An S4 class that contains the data from water monitoring for a single water quality characteristic as a single site in a single park.
}
\section{Slots}{

\describe{
\item{\code{CharacteristicName}}{The name of the water quality characteristic. Stored as a length 1 character vector.}

\item{\code{DisplayName}}{A name for the characteristic suitable for display on graphs, tables etc. Stored as a length 1 character vector.}

\item{\code{Substrate}}{The substrate of the characteristic, stored as a length 1 character vector.}

\item{\code{SampleFraction}}{The sample fraction of the characteristic. Stored as a length 1 character vector.}

\item{\code{Category}}{The category for the characteristic. Stored as a length 1 character vector.}

\item{\code{CategoryDisplay}}{The display name for the category, suitable for graphs, tables etc.}

\item{\code{Details}}{A description of the characteristic as needed. Stored as a length 1 character vector.}

\item{\code{Units}}{The units of measurement of the characteristic. Stored as a length 1 character vector.}

\item{\code{Data}}{A \code{data.frame} containing the water quality data. Should have columns representing the date, measurement, any QAQC flags etc. for each measurement.}

\item{\code{LowerPoint,UpperPoint}}{Indicates the assessment points. Values lower than the lower point are considered to have failed the assessment, 
whereas values higher than the higher point are considered failures. Can either be a length 1 numerical vector, or a \code{data.frame} 
giving the lower assessment value for each measurement. The vector method should be used for assessments that do not change, whereas the \code{data.frame} is for 
assessment that change based on the time of the year or other factors.}

\item{\code{LowerType,UpperType}}{The type of assessment indicated by the lower and upper points. A length 1 character vector.}

\item{\code{LowerDescription,UpperDescription}}{A description of the lower and upper assessments point. Stored as a length 1 character vector.}

\item{\code{AssessmentDetails}}{Additional description of the assessment point. Stored as a length 1 character vector.}
}}

