% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruency.R
\name{congruency}
\alias{congruency}
\title{Check structural and semantic congruency between WQP data and NCRNWater metadata}
\usage{
congruency(
  data_filename,
  metadata_filename,
  data_template = system.file("templates", "wqp.csv", package = "NCRNWater"),
  metadata_template = system.file("templates", "wqp_ncrnwater_metadata.csv", package =
    "NCRNWater")
)
}
\description{
`congruency()` validates whether a user-provided WQP data file and NCRNWater
metadata file are structurally compatible with each other and with the
expected NCRNWater templates. The function checks:

* that both user files exist and are CSVs
* that the **columns** in the user data and metadata match those in the
  official NCRNWater template files
* that key **value pairs** (e.g., SiteCode vs SiteCodeWQX, CharacteristicName vs DataName)
  are consistent across files
* that the user data is internally consistent with the user metadata

The function prints a human‑readable report to the console and halts with an error
}
\examples{
\dontrun{
library(NCRNWater)
my_data <- 'data/wqp.csv'
my_metadata <- 'data/metadata.csv'
NCRNWater::congruency(my_data, my_metadata)
}

}
