% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceed.R
\name{.exceed_core}
\alias{.exceed_core}
\title{Internal: core exceed implementation (do not export)}
\usage{
.exceed_core(
  object,
  parkcode = NA,
  sitecode = NA,
  charname = NA,
  category = NA,
  points = "both",
  lower = NA,
  upper = NA,
  all = FALSE,
  catsum = FALSE,
  ...
)
}
\description{
Internal: core exceed implementation (do not export)
}
\keyword{internal}
