% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_paths.R
\name{example_paths}
\alias{example_paths}
\title{Locate NCRNWater example CSVs installed with the package}
\usage{
example_paths(network = "NCRN")
}
\arguments{
\item{network}{Character. Default: "NCRN".}
}
\value{
A named list with:
  \describe{
    \item{dir}{Directory containing the example files.}
    \item{data_fp}{Absolute path to the WQP example data file.}
    \item{metadata_fp}{Absolute path to the example metadata file.}
    \item{data}{Filename (basename) of the WQP data file.}
    \item{metadata}{Filename (basename) of the metadata file.}
  }
}
\description{
Returns:
  - dir: directory under inst/extdata/<network>/
  - data_fp / metadata_fp: full paths to example CSVs
  - data / metadata: basenames parsed from those paths
}
\examples{
paths <- example_paths()
importNCRNWater(
    Dir      = paths$dir,
    Data     = paths$data,
    MetaData = paths$metadata,
    wqx      = TRUE
)

}
