% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCharacteristics.R
\name{getCharacteristics}
\alias{getCharacteristics}
\title{getCharacteristics}
\usage{
getCharacteristics(object, sitecode = NA, charname = NA)
}
\arguments{
\item{object}{One of a \code{Site} object, a \code{Park} object or a list of \code{Site} and/or \code{Park} objects.}

\item{sitecode}{One of more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the site code argument can be used to select which sites should be ssued}

\item{charname}{Name of one or more characteristics, in quotes.}
}
\value{
A list of one or more characteristic objects.
}
\description{
Retrieves one or more \code{Characteristic} objects from a \code{Site} object or a \code{list} of such objects. The \code{Site} object can be contained within a \code{Park} object.
}

