% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChars.R
\name{getChars}
\alias{getChars}
\title{getChars}
\usage{
getChars(object, parkcode = NA, sitecode = NA, charname = NA,
  category = NA)
}
\arguments{
\item{object}{One of a \code{Characteristic} object, a \code{Site} object, a \code{Park} object or a list of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Name of one or more characteristics, in quotes. Only those characteristics will be returned.}

\item{cateogory}{One or more categories of characteristics, in quotes. Only those categories of charateristics will be returned.}
}
\value{
A list of one or more \code{Characteristic} objects.
}
\description{
Retrieves one or more \code{Characteristic} objects from a \code{Site} object or a \code{list} of such objects. The \code{Site} object can be contained within a \code{Park} object.
}
