% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkNames.R
\name{getParkNames}
\alias{getParkNames}
\title{getNParkNames}
\usage{
getParkNames(object, name.class = "short")
}
\arguments{
\item{object}{Either a \code{Park} object or a \code{list} of such objects.}

\item{name.class}{Type of name to return. One of four options, in quotes.
\describe{
\item{"short"}{ The default. Returns the short name of the park}
\item{"long"}{Returns the long name of the park}
\item{"code"}{Returns the park code}
\item{"network}{Returns the network code}
}}
}
\value{
A character vector with one or more park names.
}
\description{
Retreives park names from an \code{Park} object or a \code{list} of such objects.
}

