% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getSites}
\alias{getSites}
\title{getSites}
\usage{
getSites(object, parkcode = NA, sitecode = NA)
}
\arguments{
\item{object}{Either a \code{Park} object or a \code{list} of such objects.}

\item{parkcode}{Park code of one or more parks, in quotes.}

\item{sitecode}{Site code of one or more sites, in quotes.}
}
\value{
A list of one or more site objects. If \code{parkcode} and/or \code{sitecode} are specificed then only site which match those codes will be returned. If there are no such sites then the fuciton will return \code{NULL}
}
\description{
Retrieves sites from a \code{Park} object or a \code{list} of such objects.
}
