% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWData.R
\name{getWData}
\alias{getWData}
\title{getWData}
\usage{
getWData(object, sitecode = NA, charname = NA, mindate = NA,
  maxdate = NA)
}
\arguments{
\item{object}{Either a \code{Characteristic} object a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{sitecode}{One of more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the site code argument can be used to select which sites should be ssued}

\item{charname}{Name of one or more characteristics, in quotes.}

\item{mindate}{The earliest date of data to include, in quotes. Date is in the same format as the \code{Date} in the \code{Data} slot.}

\item{maxdate}{The latest date of data to include, in quotes. Date is in the same format as the \code{Date} in the \code{Data} slot.}
}
\value{
A \code{data.frame} based on the \code{Data} slot of \code{Charactersitic} objects.
}
\description{
Gets data from \code{Characteristic} objects and formats it for ease in analysis.
}
\details{
This fuction is intended to be the primary method for accessing water data. Data from various \code{objects} (see below) is comined into a single \code{data.frame} along with comumns indicating which park/site/characteristic the data is from. 
 
   If \code{object} is a \code{Charactersitic} object \code{getWData} will return a \code{data.frame} with colums consisting of \code{Date} and \code{Value} from the \code{Data} slot as well as a column named \code{Characteristic} with the data from the {charname} slot. If \code{object} is a \code{Site} object it will return a \code{data.frame} that also includes a column named \code{Site} which comes from the \code{SiteCode} slot. If \code{object} is a \code{Park} object then a column named \code{Park} is included which comes from the \code{ParkCode} slot.
}

