% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRNWater_NCRNWaterObj_Class_def.R
\docType{class}
\name{NCRNWaterObj-class}
\alias{NCRNWaterObj-class}
\title{S4 Class Definition for the class union \code{NCRNWaterObj}}
\description{
This is an virtual class combining \code{Park}, \code{Site}, \code{Characteristic} 
and \code{list}. It is used for convenience for writing methods for the \code{NCRNWater} package.
}
