% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparTrends.R
\name{nonparTrends}
\alias{nonparTrends}
\title{nonparTrends}
\usage{
nonparTrends(object, parkcode, sitecode, charname, censored = FALSE)
}
\arguments{
\item{object}{A \code{Park} object or a \code{list} of such objects.}

\item{censored}{Either \code{TRUE} or \code{FALSE}. If FALSE (default), function runs the openair::TheilSen test with deseason = TRUE. 
If TRUE, function runs a separate NADA::cenken test for for each month in the dataset.}

\item{...}{Additional commands passed to \code{\link{getWData}} for filtering or subsetting the data.}
}
\value{
The results of a Theil-Sen analysis for censored or non-censored data with a row for month in the analysis and a column for
the month, slope, intercept and p-value (pval) from each test.
}
\description{
Uses non-parametric methods to test for trends in water quality data across years for each month. 
This function assumes that the data are composed of monthly observations that are correlated within a year. 
Note that this function is only appropriate for trends that are one-directional.
}
