% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRNWater_Park_Class_def.R
\docType{class}
\name{Park-class}
\alias{Park-class}
\title{S4 Class Definition for Park object in NCRNWater}
\description{
An S4 class that contains the data from water monitoring from a single park. Data on sites will be stored as one or more S4 objects in  each park object
}
\section{Slots}{

\describe{
\item{\code{ParkCode}}{A short code to designate the park, typically an NPS 4 letter code. Stored as a length 1 character vector.}

\item{\code{ShortName}}{A short name for the park. Stored as a length 1 character vector.}

\item{\code{LongName}}{A long, formal name for the park. Stored as a length 1 character vector.}

\item{\code{Network}}{The code for the Inventory & Monitoring network the park belongs to. Stored as a length 1 character vector.}

\item{\code{Sites}}{A list of \code{Site} objects associated with the park.}
}}

