% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbands.R
\name{waterbands}
\alias{waterbands}
\title{waterbands}
\usage{
waterbands(
  object,
  parkcode = NA,
  sitecode = NA,
  charname = NA,
  category = NA,
  year_current = NA,
  year_historic = NA,
  months = c(5:10),
  assessment = TRUE,
  param_name = NA,
  unit = NA,
  yname = NA,
  ...
)
}
\arguments{
\item{object}{Either a \code{data.frame} that is the output of \code{getWData}, a \code{Characteristic} object, a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Required if \code{object} is not a \code{data.frame}. Name(s), in quotes, of one or more \code{Characteristic}s 
whose data should be graphed.}

\item{category}{One or more categories of characteristics, in quotes. Only those categories of characteristics will be returned.}

\item{year_current}{Year that will be plotted separately.}

\item{year_historic}{First year to include in historic range calculations. Last year will be the year prior to year_current.}

\item{months}{A numeric vector corresponding to months of the year. Only data from those months will be returned.
Ranges from 1 to 12. Default is c(5:10) for May to October.}

\item{assessment}{Vector indicating if assessment lines will be marked on the graph. See details below.}

\item{param_name}{Text, defaults to \code{NA}. Used for plotly tooltips}

\item{unit}{Text, defaults to \code{NA}. Used for plotly tooltips}

\item{yname}{Text, defaults to \code{NA}. Used for y axis title}

\item{...}{Additional arguments used to select and filter data passed to \code{\link{getWData}}}
}
\value{
Creates a plot that compares current with historic ranges. If the minimum or maximum values return a tie, the most recent year is returned.
}
\description{
Produces a plot that summarizes the range of historic data and compares with current measurements.
}
\details{
The \code{assessment} argument determines if lines representing the assessment values should be drawn on the graph. 
If \code{FALSE} then no lines will be drawn. If \code{TRUE}, the default, then the upper and lower points indicated in 
\code{object}'s \code{Character} objects will be used to draw the lines. Note that if there are multiple assessment points, 
for example if different parks have different points, or if there is both an upper and lower point, they will all be drawn. 
If a \code{vector} of numbers is passed to \code{assessemnt} instead then those will serve as the assessment values and lines will 
be drawn accordingly. Note that if \code{object} is a \code{data.frame} then the only way to draw assessment lines is by passing a 
numeric \code{vector} to \code{assessment}.
}
\examples{

netnwd <- importNCRNWater(Dir = "~/Data/NETN", 
                          Data = "Water Data.csv", MetaData = "VizMetaData.csv")
parkcode = "MABI"
sitecode = "NETN_MABI_PA00"
charname = "pH"

waterbands(netnwd, parkcode = parkcode, sitecode = sitecode, charname = charname,
           year_historic = 2006, year_current = 2019, months = c(5:10), 
           assessment = TRUE)


}
