% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbox.R
\name{waterbox}
\alias{waterbox}
\title{waterbox}
\usage{
waterbox(object, groupby = "year", ...)
}
\arguments{
\item{object}{Either a \code{data.frame} that is the output of \code{getWData}, a \code{Characteristic} object a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{groupby}{Indicates how the data for the boxplot should be grouped. A text variable in quotes. Choices are:
\describe{
\item{"year"}{The default. Will prodice a boxplot for each year in the data}
\item{"month"}{ Will prodoce a boxplot for each month, combining data across years. That is all January data will be in one box, all February data in another, etc.}
}}

\item{...}{Additional arguments used to select and filter data passed to \code{\link{getWData}}}
}
\value{

}
\description{
Produces box plots from water data.
}
\details{

}

