% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterheat.R
\name{waterheat}
\alias{waterheat}
\title{waterheat}
\usage{
waterheat(object, charname, by = "year", yname = NA, xname = NA,
  labels = NA, title = NULL, webplot = FALSE, ...)
}
\arguments{
\item{object}{Required. Either a \code{data.frame} that is the output of \code{getWData}, a \code{Characteristic} object a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{charname}{Required if \code{object} is not a \code{data.frame}. Name, in quotes, of a single \code{Characteristic} whose data should be graphed.}

\item{by}{Indicates how the data for the boxplot should be grouped. A text variable in quotes. Choices are:
\describe{
\item{"month"}{ Will prodoce a boxplot for each month, combining data across years. That is all January data will be in one box, all February data in another, etc.}
\item{"site}{If more than one site is included in the input object, this will produce a different box for each site.}
}}

\item{yname}{Text, defaults to \code{NA}. A label for the y-axis. The y-label will default to whatever is chosen in \code{by}. unless overwitten by the \code{yname} argument.}

\item{xname}{Text, defaults to \code{NA}. A label for the x-axis. If left as \code{NA} and \code{by} is "month" then \code{xname} defaults to "year", if \code{by} is "Site" then \code{xname} will be "Date".}

\item{labels}{A character vector indicating the labels for the ticks on the y-axis. bars, If left as \code{NA}, and \code{by} is "month", the tick labels will be the three letter abbreviation for the month, if \code{by} is "site", and object is a \code{Site} or \code{Park} object, then the tick labels will be the site names.}

\item{title}{A title in the graph in quotes. Defauls to \code{NULL}, which indicates no title should be used.}

\item{...}{Additional arguments used to select and filter data passed to \code{\link{getWData}}}
}
\value{
Creates a heatmap
}
\description{
Produces heat maps from water data.
}

