% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterseries.R
\name{waterseries}
\alias{waterseries}
\title{waterseries}
\usage{
waterseries(object, parkcode = NA, sitecode = NA, charname, by = "none",
  points = TRUE, layers = c("points", "line"), xname = NA, yname = NA,
  labels = NA, title = NULL, webplot = FALSE, ...)
}
\arguments{
\item{object}{Either a \code{data.frame} that is the output of \code{getWData}, a \code{Characteristic} object, a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Required if \code{object} is not a \code{data.frame}. Name(s), in quotes, of one or more \code{Characteristic}s whose data should be graphed.}

\item{by}{Indicates how the data for the plot should be grouped. A text variable in quotes. Choices are:
\describe{
\item{"site}{If more than one site is included in the input object, this will produce a different line for each site.}
\item{"park"}{If more than one park is included in the input object, this will produce a different line for each park}
\item{"char"}{If more than one characteristic is included in the input object, this will produce a different line for each one} 
\item{"none}{All data will be displayed as a single ine. Used if you only have data from a single charactersitic from a single site}
}}

\item{points}{Vector indicating if assesment points will be marked on the graph. See details below.}

\item{layers}{Defaults to c("points","line") Indicates which layers you wish to see on the plot.}

\item{xname}{Text, defaults to \code{NA}. A label for the x-axis. If a \code{Characteristic}, \code{Site}, or \code{Park} object is passed to \code{object}, then the x-label will default to whatever is indicated in \code{by}.unless overwitten by the \code{xname} argument. If a \code{data.frame} is passed then the x-label will either be the text from \code{xname} or blank if \code{xname} is left as \code{NA}.}

\item{yname}{Text, defaults to \code{NA}. A label for the y-axis. If an \code{Characteristic}, \code{Site}, or \code{Park} object is passed to \code{object}, then the y-label will default to the Display Name and Units for the Charecteristic, unless overwitten by the \code{yname} argument. If a \code{data.frame} is passed then the y-label will either be the text from \code{yname} or blank if \code{yname} is left as \code{NA}.}

\item{labels}{A character vector indicating the labels for the data series, defaults to NA. If labels are provided (one for each series) they will be printed. If \code{object} is a \code{data.frame} and \code{labels} is \code{NA} then no labels will be printed. If \code{object} is a \code{Characteristic}, \code{Site}, or \code{Park} object, and \code{labels} is \code{NA} then the default will depend on the \code{by} argument. "site" willbe labled with the site name from the \code{Site}'s \code{Display Name} slot and "park" with the \code{Park}'s short name from the \code{ShortName} slot.}

\item{title}{A title in the graph in quotes. Defauls to \code{NULL}, which indicates no title should be used.}

\item{...}{Additional arguments used to select and filter data passed to \code{\link{getWData}}}
}
\value{
Creates a time series plot.
}
\description{
Produces a time series plot from water data.
}
\details{
The \code{points} argument determines if lines representing the assessment points should be drawn on the graph. If \code{FALSE} then no lines will be drawn. If \code{TRUE}, the default, then the upper and lower points indicated in \code{object}'s \code{Character} objects will be used to draw the lines. Note that if there are multiple assessemnt points, for example if diffrernt parks have different points, or if there is both an upper and lower point, they will all be drawn. If a \code{vector} of numbers is passed to \code{points} instead then those will serve as the assessment points and lines will be drawn accordingly. Note that if \code{obejct} is a \code{data.frame} then the only way to draw assessment points is by passing a numeric \code{vector} to \code{points}.
}

