% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersite_comps.R
\name{watersite_comps}
\alias{watersite_comps}
\title{watersite_comps}
\usage{
watersite_comps(
  object,
  parkcode = NA,
  sitecode = NA,
  charname = NA,
  category = NA,
  year = NA,
  months = c(5:10),
  param_name = NA,
  unit = NA,
  yname = NA,
  legend = "bottom",
  layers = c("points", "line"),
  assessment = TRUE,
  ...
)
}
\arguments{
\item{object}{Either a \code{data.frame} that is the output of \code{getWData}, a \code{Characteristic} object, a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Required if \code{object} is not a \code{data.frame}. Name(s), in quotes, of one or more \code{Characteristic}s 
whose data should be graphed.}

\item{category}{One or more categories of characteristics, in quotes. Only those categories of characteristics will be returned.}

\item{year}{Year of data that will be plotted for all sites. Must be numeric and 4 digits.}

\item{months}{A numeric vector corresponding to months of the year. Only data from those months will be returned.
Ranges from 1 to 12. Default is c(5:10) for May to October.}

\item{param_name}{Text, defaults to \code{NA}. Used for plotly tooltips}

\item{unit}{Text, defaults to \code{NA}. Used for plotly tooltips}

\item{yname}{Text, defaults to \code{NA}. Used for y axis title}

\item{legend}{a vector indicating where the legend position. Can be: "none","left","right","top","bottom" or a two element 
numeric vector.}

\item{layers}{Defaults to c("points","line") Indicates which layers you wish to see on the plot.}

\item{assessment}{Vector indicating if assessment lines will be marked on the graph. If multiple thresholds exist among the sites plotted, all will be plotted. See details below.}

\item{...}{Additional arguments used to select and filter data passed to \code{\link{getWData}}}
}
\value{
Creates a plot that compares multiple sites for a given year.
}
\description{
Produces a plot that compares multiple sites for a given year and parameter, with each site having a different color and shape.
}
\details{
The \code{assessment} argument determines if lines representing the assessment values should be drawn on the graph. 
If \code{FALSE} then no lines will be drawn. If \code{TRUE}, the default, then the upper and lower points indicated in 
\code{object}'s \code{Character} objects will be used to draw the lines. Note that if there are multiple assessment points, 
for example if different parks have different points, or if there is both an upper and lower point, they will all be drawn. 
If a \code{vector} of numbers is passed to \code{assessemnt} instead then those will serve as the assessment values and lines will 
be drawn accordingly. Note that if \code{object} is a \code{data.frame} then the only way to draw assessment lines is by passing a 
numeric \code{vector} to \code{assessment}.
}
\examples{

netnwd <- importNCRNWater(Dir = "~/Data/NETN", 
                          Data = "Water Data.csv", MetaData = "VizMetaData.csv")
parkcode = "MABI"
charname = "pH"
site_list <- getSiteInfo(netnwd, parkcode = parkcode, info = "SiteCode")

watersite_comps(netnwd, parkcode = parkcode, sitecode = site_list, charname = charname,
                year = 2019, months = c(5:10), legend = "none", 
                layers = c("points", "line"), assessment = TRUE)


}
