% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcosinor.R
\name{wcosinor}
\alias{wcosinor}
\title{wcosinor}
\usage{
wcosinor(object, ...)
}
\arguments{
\item{object}{Either a \code{Characteristic} object a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{...}{Additional commands passed to \code{\link{getWData}} for filtering or subsetting the data.}
}
\value{
The results of a cosinor analysis.
}
\description{
Does a cosinor analysis on water quality data.
}
