% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSite.R
\name{addSite}
\alias{addSite}
\title{addSite}
\usage{
addSite(park, SiteCode, SiteName, Lat, Long, Type)
}
\arguments{
\item{park}{A \code{Park} object}

\item{SiteCode}{A short code to designate the site, stored as a length 1 character \code{vector}.}

\item{SiteName}{A name for the site. Stored as a length 1 character \code{vector}.}

\item{Type}{The type of water body at the site - e.g. Stream, River, Lake, Ocean etc. Stored as a length 1 character \code{vector}.}

\item{Coordinates}{A length 2 numeric \code{vector} containing the latitude and longitude of the site in WGS84 coordinates.}
}
\value{
Returns the park object with the site added.
}
\description{
This function adds a site to an existing \code{Park} object.
}
\details{
This command can be used to add a \code{Site} object to the list of existing sites in an existing \code{Park} object. The \code{SiteCode} will be used to name the Site in the list. \code{addCharacteristic} can be used to add a charactersitic to the site.
}
