% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileNETNdata.R
\name{compileNETNdata}
\alias{compileNETNdata}
\title{compileNETNdata}
\usage{
compileNETNdata(
  path = "./Data/",
  export = TRUE,
  surface = TRUE,
  active_site = TRUE,
  active_metric = TRUE,
  restricted = TRUE,
  cleanEnv = TRUE
)
}
\arguments{
\item{path}{Quoted path to export csv files to. Defaults to "Data" folder within current project.}

\item{export}{\code{TRUE} or \code{FALSE}. Export csv files to specified path. Defaults to \code{TRUE}.}

\item{surface}{\code{TRUE} or \code{FALSE}. Return only measurements representing the stream surface or lake epilimnion. 
If \code{TRUE}, the median of the surface measurements from the top 2m of sampling are returned. Defaults to  \code{TRUE}.}

\item{active_site}{\code{TRUE} or  \code{FALSE}. If \code{TRUE} only compiles data for sites that are actively being monitored.
If \code{FALSE} compiles all sites stored in the database. Defaults to \code{TRUE}.}

\item{active_metric}{\code{TRUE} or  \code{FALSE}. If \code{TRUE} only compiles metrics that are actively collected.
If \code{FALSE} compiles all metrics stored in the database. Defaults to \code{TRUE}.}

\item{restricted}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, metrics within a site that have 4 or fewer
measurements since monitoring began in 2006 are dropped.}

\item{cleanEnv}{\code{TRUE} or  \code{FALSE}. Allows you to clean the global environment so that 
only waterDat and MD are kept. Defaults to \code{TRUE}.}
}
\value{
Returns a list of two dataframes (WaterData and MetaData) formatted for import into NCRNwater R package.Exports data to folder by default.
}
\description{
Compile flat files from NETN Access Database for R package
}
\details{
This function create flat files from NETN's water database. The function returns 2 data frames used for the 
NCRNWater R package (Water Data and Meta Data).
Currently the Water Data list element can include (1) measurements at all sampled depths (for plotting profiles) AND the 
 the median value of measurements taken within the top 2m of the water surface when \code{surface == FALSE} or 
 (2) just the median values of  surface measurements within the top 2m of the water surface  \code{surface == TRUE}.
The latter is to enable examining patterns in the stream surface or lake epilimnion over time and is currently (as of 11/1/2019) the data working 
in the NCRNwater package.
}
\examples{
# compile NETN water data for surface measurements and active metrics only,
# and export to C:/Data.
compileNETNdata(path = "C:/Data/", export = TRUE, surface = TRUE, active = TRUE, 
  restricted = TRUE)

}
