% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceed.R
\name{exceed}
\alias{exceed}
\title{exceed}
\usage{
exceed(
  object,
  parkcode = NA,
  sitecode = NA,
  charname = NA,
  category = NA,
  points = "both",
  lower = NA,
  upper = NA,
  all = F,
  catsum = F,
  ...
)
}
\arguments{
\item{object}{One of a \code{Characteristic} object, a \code{Site} object, a \code{Park} object or a list of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Name of one or more characteristics, in quotes. Only those characteristics will be returned.}

\item{category}{One or more categories of characteristics, in quotes. Only those categories of characteristics will be returned.}

\item{points}{What assessment points to use when determining if there are exceedances. There are three options
\describe{
\item{"lower"}{Only assess the data against the lower reference point}
\item{"upper"}{Only assess the data against the upper reference point}
\item{"both"}{The default. Assess the data against both reference points}
}}

\item{lower, upper}{The lower and uppers assessment points. Either a number specified by the user, or if \code{NA}, the default,
the assessment point is determined by the \code{LowerPoint} and \code{UpperPoint} slots in the \code{Characteristic} objects.}

\item{all}{Logical, defaults to \code{FALSE}. Not used when \code{object} is a \code{data.frame}. If \code{all} is \code{FALSE} 
characteristics without upper or lower references points will not be included in the results. If \code{all} is \code{TRUE} such characteristics will be included.}

\item{catsum}{Logical, defaults to \code{False}. When true summarizes by category rather than a characteristic. Not used if 
\code{object} is a \code{data.frame}}

\item{...}{Additional commands passed to \code{\link{getWData}} for filtering or subsetting the data.}
}
\description{
Determines if water data contains observations that exceed a upper and/or lower assessment points.
}
