% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceed.R
\name{exceed}
\alias{exceed}
\title{exceed}
\usage{
exceed(object, parkcode = NA, sitecode = NA, charname = NA,
  points = "both", lower = NA, upper = NA, all = F, ...)
}
\arguments{
\item{object}{One of a \code{Characteristic} object, a \code{Site} object, a \code{Park} object or a list of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Name of one or more characteristics, in quotes.}

\item{points}{What assessemnt points to use when determining if there are exceedences. There are three options
\describe{
\item{"lower"}{Only assess the data against the lower reference point}
\item{"upper"}{Only assess the data against the upper reference point}
\item{"both"}{The default. Assess the data against both reference points}
}}

\item{lower, upper}{The lower and uppers assessment points. Either a number specified by the user, or if \code{NA}, the default, the assessment point is determined by the \code{LowerPoint} and \code{UpperPoint} slots in the \code{Characteristic} objects.}

\item{all}{Logical, defaults to \code{FALSE}. Not used when \code{object} is a \code{data.frame}. If \code{all} is \code{FALSE} characteristics without uppoer or lower references points will not be incldued in the reults. If \code{all} is \code{TRUE} such characteristics will be included.}
}
\description{
Determines if water data contains observtions that exceed a upper and/or lower assessment points.
}
