% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCharInfo.R
\name{getCharInfo}
\alias{getCharInfo}
\title{getCharInfo}
\usage{
getCharInfo(
  object,
  parkcode = NA,
  sitecode = NA,
  charname = NA,
  category = NA,
  info = NA
)
}
\arguments{
\item{object}{Either a \code{Characteristic} object or a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Name of one or more characteristics, in quotes. Only those characteristics will be returned.}

\item{category}{One or more categories of characteristics, in quotes. Only those categories of charateristics will be returned.}

\item{info}{Type of information to return. One of several options, in quotes.
\describe{
\item{"CharName"}{The name of the characteristic.}
\item{"DisplayName"}{The display name of the characteristic.}
\item{"Substrate"}{"The substrate of the characteristic.}
\item{"SampleFraction}{SampleFraction The sample fraction of the characteristic. Stored as a length 1 character vector.}
\item{"Category"}{The category for the charactersitic. Stored as a length 1 character vector.}
\item{"CategoryDisplay}{The display name for the category of the characteristic.}
\item{"Details"}{A description of the characteristic as needed. Stored as a length 1 character vector.}
\item{"Units"}{The units of measurement of the characteristic. Stored as a length 1 character vector.}
\item{"Data'}{A \code{data.frame} containing the water quality data. Should have cloumns representing the date, measurment, any QAQC flags etc. for each measurement.}
\item{"LowerPoint","UpperPoint}{Indicates the assessment points. Values lower than the lower point are considered to have failed the assessement, whereas values higher than the higher point are considered failures.}
\item{"LowerType","UpperType"}{The type of assessment indicated by the lower and upper points.}
\item{"LowerDescription","UpperDescription"}{A description of the lowerand upper assesements point.}
\item{"AssessmentDetails"}{Additional description of the assessement point.}
\item{"SiteCode"}{Returns the site code.}
\item{"SiteName"}{Returns the name of the site.}
\item{"coords"}{Returns the latitude and longitude of the site as a length 2 vector.}
\item{"type"}{Returns the type of the site.}
\item{"ParkCode"}{Returns the park code for the park the site is in.}
\item{"ParkShortName"}{ The default. Returns the short name of the park the site is in.}
\item{"ParkLongName"}{Returns the long name of the park the site is in.}
\item{"Network}{Returns the network code for the network the site is in.}
}}
}
\value{
Either a vector or a list with information for each character. Only the data option returns a list.
}
\description{
Retreives the metadata from a \code{Site} object or a \code{list} of such objects.
}
