% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteInfo.R
\name{getSiteInfo}
\alias{getSiteInfo}
\title{getSiteInfo}
\usage{
getSiteInfo(object, sitecode = NA, info)
}
\arguments{
\item{object}{Either a \code{Park} object or a \code{site} object or a \code{list} of such objects.}

\item{info}{Type of infromation to return. One of four options, in quotes.
\describe{
\item{"code"}{ The default. Returns the site code}
\item{"name"}{Returns the name of the site}
\item{"coords"}{Returns the latitude and longitude of the site as a length 2 vector}
\item{"type"}{Returns the type of the site}
}}
}
\value{
A character vector with one or more park names.
}
\description{
Retreives the metadata from a \code{Site} object or a \code{list} of such objects.
}

