% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getSites}
\alias{getSites}
\title{getSites}
\usage{
getSites(object, sitecode = NA)
}
\arguments{
\item{object}{Either a \code{Park} object or a \code{list} of such objects.}

\item{sitecode}{Name of one or more sites, in quotes.}
}
\value{
A list of one or more site objects.
}
\description{
Retrieves sites from a \code{Park} object or a \code{list} of such objects.
}

