% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNCRNWater.R
\name{importNCRNWater}
\alias{importNCRNWater}
\title{importNCRNWater}
\usage{
importNCRNWater(Dir, Data = "Water Data.csv", MetaData = "MetaData.csv")
}
\arguments{
\item{Dir}{The directory where the data is found}

\item{Data}{The data file. Defaults to "Water Data.csv"}

\item{MetaData}{The metadata file. Defaults to "MetaData.csv"}
}
\value{
Returns \code{Park} objects, one for each park, as a \code{list}.
}
\description{
This function imports data from a .csv files exporterd from NPStoret and saves it as \code{Park} objects.
}
\examples{
ncrnwd<-importNCRNWater(Dir = "./Data/NCRN", Data = "Water Data.csv", MetaData = "VizMetaData.csv")

}
