% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNCRNWater.R
\name{importNCRNWater}
\alias{importNCRNWater}
\title{importNCRNWater}
\usage{
importNCRNWater(
  Dir,
  Data = "Water Data.csv",
  MetaData = "MetaData.csv",
  wqx = FALSE
)
}
\arguments{
\item{Dir}{The directory where the data is found}

\item{Data}{The data file. Defaults to "Water Data.csv"}

\item{MetaData}{The metadata file. Defaults to "MetaData.csv"}

\item{wqx}{Indicates if the data file is in wqx format, defaults to \code{FALSE}.}
}
\value{
Returns \code{Park} objects, one for each park, as a \code{list}.
}
\description{
This function imports data from a .csv files exported from NPStoret and saves it as \code{Park} objects.
}
\examples{
ncrnwd<-importNCRNWater(Dir = "./Data/NCRN", Data = "Water Data.csv", MetaData = "VizMetaData.csv")

}
