% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNCRNWater.R
\name{importNCRNWater}
\alias{importNCRNWater}
\title{importNCRNWater}
\usage{
importNCRNWater(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found}
}
\value{
Returns 11 \code{Park} objects, one for each park, as a \code{list}.
}
\description{
This function imports data from a .csv files exporterd from NPStoret and saves it as \code{Park} objects.
}

