% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySplashService.R
\name{resolve_splash}
\alias{resolve_splash}
\title{Resolve splash (loading) text and images from config}
\usage{
resolve_splash(cfg)
}
\description{
Resolve splash (loading) text and images from config
}
\keyword{internal}
