% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbox.R
\name{waterbox}
\alias{waterbox}
\title{waterbox}
\usage{
waterbox(object, parkcode = NA, sitecode = NA, charname = NA,
  category = NA, by = "year", assessment = TRUE, yname = NA,
  xname = NA, labels = NA, title = NULL, assesscolor = "red",
  outliercolor = "blue", sizes = c(2, 0.5, 1), webplot = FALSE, ...)
}
\arguments{
\item{object}{Either a \code{data.frame} that is the output of \code{getWData}, a \code{Characteristic} 
object a \code{Site} object, a \code{Park} object or a \code{list} of such objects.}

\item{parkcode}{One or more parkcodes, in quotes.  If \code{object} is a \code{Park} object or a \code{list} of \code{Park} objects, then the \code{parkcode} argument can be used to select which sites should be used.}

\item{sitecode}{One or more sitecodes, in quotes. If \code{object} is a \code{Park} object or a \code{list} of \code{Park} or \code{Site} objects, then the \code{sitcode} argument can be used to select which sites should be used.}

\item{charname}{Name, in quotes, #' of a single \code{Characteristic} whose data should be graphed. This is required unless 
\code{category} is specified or \code{object} is a \code{data.frame}.}

\item{category}{Name, in quotes, #' of a single catogory of characteristics whose data should be graphed. This is required unless 
\code{charname} is specified or \code{object} is a \code{data.frame}.}

\item{by}{Indicates how the data for the boxplot should be grouped. A character variable in quotes. Choices are:
\describe{
\item{"year"}{The default. Will produce a boxplot for each year in the data}
\item{"month"}{ Will produce a boxplot for each month, combining data across years. That is all January data will be in one box, 
all February data in another, etc.}
\item{"site}{If more than one site is included in the input object, this will produce a different box for each site.}
\item{"park"}{If more than one park is included int he input object, this will produce a different box for each park.}
\item{"char"}{If a single category of charactersitics is chosen, this will produce a different box for each of
the component characteristics.}
}}

\item{assessment}{Vector indicating if assessment points will be maked on the graph. See details below.}

\item{yname}{Text, defaults to \code{NA}. A label for the y-axis. If an \code{Characteristic}, \code{Site}, or \code{Park} object 
is passed to \code{object}, then the y-label will default to the Display Name and Units for the Charecteristic, unless overwitten by
 the \code{yname} argument. If a \code{data.frame} is passed then the y-label will either be the text from \code{yname} or blank 
 if \code{yname} is left as \code{NA}.}

\item{xname}{Text, defaults to \code{NA}. A label for the x-axis. If a \code{Characteristic}, \code{Site}, or \code{Park} object is 
passed to \code{object}, then the x-label will default to whatever is indicated in \code{by}.unless overwitten by the \code{xname} 
argument. If a \code{data.frame} is passed then the x-label will either be the text from \code{xname} or blank if \code{xname} is 
left as \code{NA}.}

\item{labels}{A character vector indicating the labels for the bars, defaults to NA. If labels are provided (one for each bar) they 
will be printed. If \code{object} is a \code{data.frame} and \code{labels} is \code{NA} then no labels will be printed. 
If \code{object} is a \code{Characteristic}, \code{Site}, or \code{Park} object, and \code{labels} is \code{NA} then the default will 
depend on the \code{gropby} argument. "year" will be labeled with 4 digit years, "month" with 3 letter months (Jan,Feb, etc), "site" 
with the site name from the \code{Site}'s \code{Display Name} slot "park" with the \code{Park}'s short name from the 
\code{ShortName} slot and "char" will use the names of the characteristics.}

\item{title}{A title for the graph in quotes. Defaults to \code{NULL}, which indicates no title should be used.}

\item{assesscolor}{a length one character vector with the color for the assessment lines.}

\item{outliercolor}{a length one character vector with the color for the outlier points.}

\item{sizes}{a length 3 numeric vector with the sizes for the outlier points, lines of the boxplot and assessment lines.}

\item{webplot}{If TRUE, the plot is produced using ggploty from the ploty package. Will produce a html plot with interactive features.}

\item{...}{Additional arguments used to select and filter data passed to \code{\link{getWData}}}
}
\value{
Creates a boxplot
}
\description{
Produces box plots from water data.
}
\details{
The \code{assessment} argument determines if lines representing the assessment points should be drawn on the graph. 
If \code{FALSE} then no lines will be drawn. If \code{TRUE}, the default, then the upper and lower assessment points indicated 
in \code{object}'s \code{Character} objects will be used to draw the lines. Note that if there are multiple assessemnt points, 
for example if diffrerent sites have different points, or if there is both an upper and lower point, they will all be drawn. If 
a \code{vector} of numbers is passed to \code{assessment} instead then those will serve as the assessment points and lines will 
be drawn accordingly. Note that if \code{obejct} is a \code{data.frame} then the only way to draw assessment points is by passing 
a \code{numeric vector} to \code{assessment}.
}
